requirements.txt
# MTA Train Display - Python Dependencies

# Core dependencies
requests>=2.28.0
Pillow>=9.0.0

# Protocol Buffers for GTFS-RT parsing
protobuf>=3.20.0
gtfs-realtime-bindings>=0.0.4

# For Raspberry Pi LED matrix control (choose one):
# Option 1: Adafruit rgbmatrix library (recommended for new projects)
# Note: Requires compilation on RPi, see README for instructions
# rgbmatrix>=1.9.2

# Option 2: rpi-rgb-led-matrix (alternative)
# install via: git clone https://github.com/hzeller/rpi-rgb-led-matrix.git

# Optional dependencies for enhanced functionality
# numpy>=1.20.0  # For advanced image processing if needed
# opencv-python>=4.5.0  # For video processing if needed

# Development dependencies
# pytest>=6.0.0  # For testing
# black>=21.0  # For code formatting
